Bindable system
Consists of Bindable parent class, which used only to return *'s to Device and Context, and function Bind.
It's children are all pipeline stages, where we set all descriptors and Device->CreateSomething in constructor.
Vertex buffer have template constructor so it can take any structure to create a buffer.
VertexShader have public getter for Bytecodeblob so we can use it for InputLayout stage.

Constant buffer class is a template class, with two constructors. One takes only Graphics& gfx, and another also
takes constant data. It have Update function, which Maps buffer to subresource, writes constant data to it,  and releases.
Also it have two children PixelConstBuffer and VertexConstBuff, which bind buffers to corresponding stages.

TransformConstantBuffer takes Shape parent to get previous state of constant buffers.
It's bind function have side effect where it uses ConstantBuffer's Update  function to multiply
Transform * Projection(this one is in Graphics class, creates projection, in that order)

Class shape is a polymorphic class, which have AddBind and AddIndexBuffer, second needs to be separated
because we need index count, so we create a separate pointer to index buffer.
Function BindAndDraw binds every member of pipeline and calls DrawIndexed, which is in Graphics class
It have a GetTransform virtual fuction which creates transform, virtual Update(float dt), which updates constant buffers.